refactor to remove the check integrations step on nodejs-worker - now everything is going on in job-generator
