+++ backend/src/bin/nodejs-worker.ts
@@ -10,7 +10,7 @@ import { NodeWorkerMessageType } from '../serverless/types/workerTypes'
 import { sendNodeWorkerMessage } from '../serverless/utils/nodeWorkerSQS'
 import { NodeWorkerMessageBase } from '../types/mq/nodeWorkerMessageBase'
 import { deleteMessage, receiveMessage, sendMessage } from '../utils/sqs'
-import { processIntegration, processIntegrationCheck, processWebhook } from './worker/integrations'
+import { processIntegration, processWebhook } from './worker/integrations'
 
 /* eslint-disable no-constant-condition */
 
@@ -158,9 +158,6 @@ async function handleMessages() {
         let processFunction: (msg: NodeWorkerMessageBase, logger?: Logger) => Promise<void>
 
         switch (msg.type) {
-          case NodeWorkerMessageType.INTEGRATION_CHECK:
-            processFunction = processIntegrationCheck
-            break
           case NodeWorkerMessageType.INTEGRATION_PROCESS:
             processFunction = processIntegration
             break
