+++ backend/src/serverless/integrations/services/integrationServiceBase.ts
@@ -2,7 +2,7 @@ import { SuperfaceClient } from '@superfaceai/one-sdk'
 import moment from 'moment'
 import crypto from 'crypto'
 import { getServiceChildLogger } from '@crowd/logging'
-import { IntegrationType } from '@crowd/types'
+import { IntegrationRunState, IntegrationType } from '@crowd/types'
 import { IRepositoryOptions } from '../../../database/repositories/IRepositoryOptions'
 import {
   IIntegrationStream,
@@ -16,7 +16,6 @@ import { IS_TEST_ENV } from '../../../conf'
 import { sendNodeWorkerMessage } from '../../utils/nodeWorkerSQS'
 import { NodeWorkerIntegrationProcessMessage } from '../../../types/mq/nodeWorkerIntegrationProcessMessage'
 import IntegrationRunRepository from '../../../database/repositories/integrationRunRepository'
-import { IntegrationRunState } from '../../../types/integrationRunTypes'
 
 const logger = getServiceChildLogger('integrationService')
